Rename and move timeline to vertical timeline
