+++ website/src/app/components/VerticalTimeline/VerticalTimeline.css
